requirements.txt
psycopg2-binary>=2.9
confluent-kafka>=2.3
pytest>=7.0
streamlit>=1.30
pandas>=2.0
plotly>=5.18
